working copy: add `check_out()` function to the backend trait
This includes documenting the new function and the other types moved
to the `working_copy` module.
