+++ lib/tests/test_local_working_copy_sparse.rs
@@ -13,11 +13,11 @@
 // limitations under the License.
 
 use itertools::Itertools;
-use jj_lib::local_working_copy::{CheckoutStats, LocalWorkingCopy};
+use jj_lib::local_working_copy::LocalWorkingCopy;
 use jj_lib::matchers::EverythingMatcher;
 use jj_lib::repo::Repo;
 use jj_lib::repo_path::RepoPath;
-use jj_lib::working_copy::WorkingCopy;
+use jj_lib::working_copy::{CheckoutStats, WorkingCopy};
 use testutils::{create_tree, TestWorkspace};
 
 #[test]
